+++ plg_content_apigrabber/services/provider.php
@@ -1,0 +1,40 @@
+<?php
+/**
+ * @package     plg_content_apigrabber
+ * @subpackage  Services
+ */
+
+defined('_JEXEC') or die;
+
+use Joomla\CMS\Extension\PluginInterface;
+use Joomla\CMS\Factory;
+use Joomla\CMS\Plugin\PluginHelper;
+use Joomla\DI\Container;
+use Joomla\DI\ServiceProviderInterface;
+use Joomla\Event\DispatcherInterface;
+use Nickpsal\Plugin\Content\ApiGrabber\Extension\ApiGrabber;
+
+return new class () implements ServiceProviderInterface {
+    /**
+     * Registers the service provider with a DI container.
+     *
+     * @param   Container  $container  The DI container.
+     *
+     * @return  void
+     */
+    public function register(Container $container): void
+    {
+        $container->set(
+            PluginInterface::class,
+            function (Container $container) {
+                $plugin = new ApiGrabber(
+                    $container->get(DispatcherInterface::class),
+                    (array) PluginHelper::getPlugin('content', 'apigrabber')
+                );
+                $plugin->setApplication(Factory::getApplication());
+
+                return $plugin;
+            }
+        );
+    }
+};